+++ test/unit/fuzz-tools.test.js
@@ -369,6 +369,7 @@ import * as koosSchwannomaV389 from '../../lib/koos-schwannoma-v389.js';
 import * as knospAdenomaV390 from '../../lib/knosp-adenoma-v390.js';
 import * as hardyAdenomaV391 from '../../lib/hardy-adenoma-v391.js';
 import * as hillFlapValveV392 from '../../lib/hill-flap-valve-v392.js';
+import * as laurenGastricV393 from '../../lib/lauren-gastric-v393.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -683,6 +684,7 @@ const MODULES = {
   'knosp-adenoma-v390.js': knospAdenomaV390,
   'hardy-adenoma-v391.js': hardyAdenomaV391,
   'hill-flap-valve-v392.js': hillFlapValveV392,
+  'lauren-gastric-v393.js': laurenGastricV393,
   'tb-testing.js': tbTesting,
 };